% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnesHook.R
\name{agnesHook}
\alias{agnesHook}
\title{AGglomerative NESting}
\usage{
agnesHook(this_dist, k)
}
\arguments{
\item{this_dist}{distance matrix}

\item{k}{scalar indicating number of clusters to cut tree into}
}
\value{
clustering assignment from agnes
}
\description{
A hook function for using agnes clustering as a custom
algorithm in ConsensusClusterPlus.
}
\details{
The hierarchical structure returned using \code{agnes} is equivalent
to using \code{hclust}.
}
\author{
Derek Chiu
}

