% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combineAlgs.R
\name{combineAlgs}
\alias{combineAlgs}
\title{Combine the consensus results from all algorithms}
\usage{
combineAlgs(..., res.CCP, k, element, alg.names = NULL)
}
\arguments{
\item{...}{any number of objects outputted from \code{ConClust}}

\item{res.CCP}{an object outputted from \code{ConClustPlus}}

\item{k}{desired number of clusters}

\item{element}{the element to extract. This is either "matrix" or "class" for
consensus matrix or consensus classes, respectively.}

\item{alg.names}{optional. Supply a vector of names for the algorithms.}
}
\value{
Either a list of all consensus matrices or a data frame
showing all the consensus classes
}
\description{
The function can output either the consensus matrices
or consensus classes for all algorithms
}
\author{
Derek Chiu
}

