% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConClust.R
\name{ConClust}
\alias{ConClust}
\title{Consensus clustering over samples and algorithms}
\usage{
ConClust(x, k, pItem = 0.8, reps = 1000, file = NULL)
}
\arguments{
\item{x}{data matrix; genes are rows and samples are columns}

\item{k}{number of clusters requested}

\item{pItem}{proportion of items to be used in subsampling within an algorithm}

\item{reps}{number of subsamples}

\item{file}{directory where returned object will be saved (at each iteration).
No output file is saved if \code{file} is \code{NULL}.}
}
\value{
An array of dimension \code{nrow(x)} by \code{reps} by \code{length(methods)}
Each slice of the array is a matrix showing consensus clustering results for
algorithms (method). The matrices have a row for each sample, and a column for each
subsample. Each entry represents a class membership.
}
\description{
Generates multiple runs for consensus clustering among replicated subsamples
of a dataset as well as across different clustering algorithms.
}
\author{
Derek Chiu, Aline Talhouk
}

