% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConClustPlus.R
\name{ConClustPlus}
\alias{ConClustPlus}
\title{Run a series of ConsensusClusterPlus algorithms}
\usage{
ConClustPlus(dat, k = 3, reps = 1000, pItem = 0.8, dir = NULL,
  seed = 123, min.sd = 1, ...)
}
\arguments{
\item{dat}{data matrix; columns are samples and rows are genes/features}

\item{k}{maximum number of clusters to compute consensus results}

\item{reps}{number of subsamples}

\item{pItem}{proportion of features to use in each subsample}

\item{dir}{directory to save results}

\item{seed}{random seed to maintain reproducible results}

\item{min.sd}{minimum standard deviation across each gene}

\item{...}{additional arguments to \code{ConsensusClusterPlus}}
}
\value{
A list of outputs from \code{ConsensusClusterPlus}; each element
is for a different algorithm.
}
\description{
Runs consensus clustering for multiple algorithms and saves the output
}
\examples{
set.seed(23)
ConClustPlus(matrix(rnorm(100), nrow = 10), k = 3, reps = 10, pItem = 0.9)
}
\author{
Derek Chiu
}

