% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityMatrix.R
\name{connectivityMatrix}
\alias{connectivityMatrix}
\title{Connectivity matrix}
\usage{
connectivityMatrix(clusters)
}
\arguments{
\item{clusters}{vector of cluster memberships}
}
\value{
A \code{length(clusters)} by \code{length(clusters)} symmetric
connectivity matrix.
}
\description{
Given a vector of cluster memberships, the connectivity matrix is computed.
}
\details{
The connectivity matrix has a 1 if both samples are in the same cluster,
and 0 otherwise.
}
\author{
Derek Chiu
}

