% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_combine.R
\name{consensus_combine}
\alias{consensus_combine}
\title{Combine consensus results}
\usage{
consensus_combine(..., res.CCP, k, element = c("matrix", "class"),
  alg.names = NULL)
}
\arguments{
\item{...}{any number of objects outputted from \code{\link{ConClust}}}

\item{res.CCP}{an object outputted from \code{\link{ConClustPlus}}}

\item{k}{desired number of clusters}

\item{element}{either "matrix" or "class" to extract the consensus matrix or
consensus class, respectively.}

\item{alg.names}{optional. Supply a vector of names for the algorithms.}
}
\value{
Either a list of all consensus matrices or a data frame showing all
  the consensus classes
}
\description{
Combine results from ConClust and ConClustPlus and output either the
consensus matrices or consensus classes for all algorithms from both objects.
}
\examples{
set.seed(911)
x <- matrix(rnorm(1000), nrow = 10)
CC1 <- ConClust(x, k = 4, reps = 10, method = "apEucl", save = FALSE)
CC2 <- ConClust(x, k = 4, reps = 10, method = "gmmBIC", save = FALSE)
CCP <- ConClustPlus(x, k = 4, reps = 10, save = FALSE)
y1 <- combineAlgs(CC1, CC2, res.CCP = CCP, k = 4, element = "matrix")
str(y1)
y2 <- combineAlgs(CC1, CC2, res.CCP = CCP, k = 4, element = "class")
str(y2)
}
\author{
Derek Chiu
}

