% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_summary.R
\name{consensus_summary}
\alias{consensus_summary}
\title{Consensus summary}
\usage{
consensus_summary(res, k, dir = NULL, file = "results_CC")
}
\arguments{
\item{res}{output from \code{ConClust}}

\item{k}{number of clusters}

\item{dir}{directory to save output}

\item{file}{file name of the written object}
}
\value{
A list with summaries for each algorithm. Each algorithm has a list
  with two elements: consensus_matrix and consensus_class
}
\description{
Given an object from \code{\link{ConClust}}, returns a list of consensus
matrices and consensus classes for each clustering algorithm.
}
\author{
Derek Chiu
}
\seealso{
Other consensus functions: \code{\link{consensus_class}},
  \code{\link{consensus_confmat}},
  \code{\link{consensus_matrix}}
}

