% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClass.R
\name{consensusClass}
\alias{consensusClass}
\title{Obtain consensus classes}
\usage{
consensusClass(x, k, method = "average", names = NULL)
}
\arguments{
\item{x}{a consensus matrix}

\item{k}{desired number of clusters}

\item{method}{linkage type for hierarchical clustering}

\item{names}{vector of names for each sample. Default is \code{NULL},
so that the names are obtained from the dendrogram labels.}
}
\value{
assignments for the consensus class
}
\description{
Performs hierarchical clustering on a consensus matrix to obtain
consensus class labels.
}
\author{
Derek Chiu
}

