% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusConfmat.R
\name{consensusConfmat}
\alias{consensusConfmat}
\title{Confusion matrix between consensus clusters and true classes}
\usage{
consensusConfmat(clust, cl.true, k, pred.lab = "Prediction",
  ref.lab = "Reference")
}
\arguments{
\item{clust}{vector of cluster assignments}

\item{cl.true}{true class labels}

\item{k}{number of clusters}

\item{pred.lab}{label for the predicted clusters}

\item{ref.lab}{label for the reference classes}
}
\value{
A confusion matrix with the predicted cluster 
assignments compared to the reference true class labels.
}
\description{
Combines the cluster assignments from multiple algorithms, then 
compares to the true class partitions.
}
\author{
Derek Chiu
}

