% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusMatrix.R
\name{consensusMatrix}
\alias{consensusMatrix}
\title{Consensus matrix}
\usage{
consensusMatrix(dat)
}
\arguments{
\item{dat}{data matrix}
}
\value{
a consensus matrix
}
\description{
Returns the consensus matrix given a data matrix
}
\details{
Input matrix has rows as samples, columns as replicates.
}
\author{
Derek Chiu
}

