% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSummaries.R
\name{consensusSummaries}
\alias{consensusSummaries}
\title{Compute consensus matrices and classes from ConClust output}
\usage{
consensusSummaries(res, k, dir = NULL, fileName = "results_CC")
}
\arguments{
\item{res}{output result from ConClust}

\item{k}{desired number of clusters}

\item{dir}{directory to save output}

\item{fileName}{file name of the written object}
}
\value{
A list with summaries for each algorithm. Each algorithm has a list with
two elements: consensusMatrix and consensusClass.
}
\description{
Given an output from ConClust, returns a list of consensus matrices and classes
for each algorithm.
}
\author{
Derek Chiu
}

