% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosimilarity_matrix.R
\name{cosimilarity_matrix}
\alias{cosimilarity_matrix}
\title{Returns cosine cosimilarity matrix for the terms generated by conclust}
\usage{
cosimilarity_matrix(x)
}
\arguments{
\item{x}{output from \code{\link[=conclust]{conclust()}}}
}
\description{
A function that extracts the cosimilarity matrix for
terms generated by \code{\link[=conclust]{conclust()}}
}
\keyword{conclust}
