% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{dataPrep}
\alias{dataPrep}
\title{Prepare data for consensus clustering}
\usage{
dataPrep(data, min.sd = 1)
}
\arguments{
\item{data}{data matrix. Columns are samples and rows are genes/features.}

\item{min.sd}{minimum standard deviation threshold. See details.}
}
\value{
dataset prepared for usage in \code{ConClust}
}
\description{
Remove variables with low signal and scale before consensus clustering
(using \code{ConClust})
}
\details{
The \code{min.sd} argument is used to filter the feature space for only highly variable
features. Only features with a standard deviation across all samples greater than
\code{min.sd} will be used.
}

