% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dianaHook.R
\name{dianaHook}
\alias{dianaHook}
\title{DIvisive ANAlysis Clustering}
\usage{
dianaHook(this_dist, k)
}
\arguments{
\item{this_dist}{distance matrix}

\item{k}{scalar indicating number of clusters to cut tree into}
}
\value{
clustering assignment from diana
}
\description{
A hook function for using diana clustering as a custom
algorithm in ConsensusClusterPlus.
}
\author{
Derek Chiu
}

