% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastText_eng_sample}
\alias{fastText_eng_sample}
\title{Sample similarity matrix generated from the pre-trained English fastText model}
\format{
A 2,000 row and 2,000 column cosine similarity matrix
}
\usage{
fastText_eng_sample
}
\description{
This similarity matrix has 2,000 rows and 2,000 columns. Each row
has numeric values that are the cosine similarity between that word
and the word of each corresponding column
}
\examples{
\donttest{
data(fastText_eng_sample)
}
}
\references{
P. Bojanowski*, E. Grave*, A. Joulin, T. Mikolov,
Enriching Word Vectors with Subword Information
(\href{https://arxiv.org/abs/1607.04606}{arxiv})
}
\keyword{data}
