% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{hooks}
\alias{agnes_hook}
\alias{diana_hook}
\alias{hooks}
\alias{mi_hook}
\title{Hook Functions}
\usage{
agnes_hook(d, k)

diana_hook(d, k)

mi_hook(x)
}
\arguments{
\item{d}{distance matrix}

\item{k}{scalar indicating number of clusters to cut tree into}

\item{x}{data matrix}
}
\value{
clustering assignment from \code{agnes} or \code{diana}; or for
  \code{mi_hook}, the mutual information distance matrix
}
\description{
Custom hook functions for AGglomerative NESting and DIvisive ANAlysis
clustering algorithms and Mutual Information distance in
ConsensusClusterPlus.
}
\details{
The hierarchical structure returned using \code{agnes} is equivalent to using
\code{\link{hclust}}.
}
\author{
Derek Chiu
}

