% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicatorMatrix.R
\name{indicatorMatrix}
\alias{indicatorMatrix}
\title{Indicator matrix}
\usage{
indicatorMatrix(clusters)
}
\arguments{
\item{clusters}{vector of cluster memberships}
}
\value{
A \code{length(clusters)} by \code{length(clusters)} symmetric
indicator matrix.
}
\description{
Given a vector of cluster memberships, the indicator matrix is computed.
}
\details{
The indicator matrix has a 1 if both samples were selected to be used in
a subsample of a consensus clustering algorithm, and 0 otherwise.
}
\author{
Derek Chiu
}

