% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Cluster validation indices}
\usage{
indices(dat, cl.mat, cons.mat, alg.names = NULL)
}
\arguments{
\item{dat}{data matrix with samples as rows, genes as columns}

\item{cl.mat}{matrix of cluster assignments. Each row is an assignment for
a different algorithm.}

\item{cons.mat}{A list of consensus matrices, one for each algorithm.}

\item{alg.names}{names for labeling the algorithms}
}
\value{
A data frame of internal evaluation indices in each column, with
algorithms in each row
}
\description{
Compute internal evaluation indices for a set of algorithms.
}
\author{
Derek Chiu
}

