% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnImpute.R
\name{knnImpute}
\alias{knnImpute}
\title{Function to any missing clustering (due to resampling using K nearest neighbour)}
\usage{
knnImpute(res, data = tdat, seed = 123456)
}
\arguments{
\item{res}{is a vector representing a column of the output from ConClust or ConClustPlus}
}
\description{
Function to any missing clustering (due to resampling using K nearest neighbour)
}
\author{
Aline Talhouk
}

