% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMIdist.R
\name{myMIdist}
\alias{myMIdist}
\title{Mutual Information}
\usage{
myMIdist(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
mutual information distance matrix
}
\description{
A hook function for using mutual information as a custom
distance metric in ConsensusClusterPlus.
}
\author{
Derek Chiu
}

