% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PAC.R
\name{PAC}
\alias{PAC}
\title{Proportion of Ambiguous Clusters}
\usage{
PAC(cons.mat, lower = 0, upper = 1)
}
\arguments{
\item{cons.mat}{consensus matrix. Should be symmetric and values between 0 and 1.}

\item{lower}{the lower bound that determines what is ambiguous}

\item{upper}{the upper bound that determines what is ambiguous}
}
\value{
the PAC is a score used in clustering performance. The lower it is the
better, because we want minimal ambiguity amongst the consensus.
}
\description{
Given a consensus matrix, returns the proportion of ambiguous clusters (PAC).
This is a robust way to assess clustering performance.
}
\details{
Since a consensus matrix is symmetric, we only look at its lower (or upper)
triangular matrix. The proportion of entries strictly between \code{lower} and
\code{upper} is the PAC. In a perfect clustering, the consensus matrix would
consist of only 0s and 1s, and the PAC assessed on the (0, 1) interval would
have a perfect score of 0. Using a (0.1, 0.9) interval for defining ambiguity
is common as well.
}
\author{
Derek Chiu
}

