% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcNormal.R
\name{pcNormal}
\alias{pcNormal}
\title{Generate null distributions on empirical gene-gene correlations}
\usage{
pcNormal(data, nSim = 50)
}
\arguments{
\item{data}{data matrix}

\item{nSim}{The number of simulated datasets to generate}
}
\value{
A list of length \code{nSim}. Each element is a pcNormal simulated
matrix.
}
\description{
Using a principal component constructed from the sample space,
we simulate null distributions with univariate Normal distributions.
}
\examples{
A <- matrix(rnorm(300), nrow = 20)
pcDats <- pcNormal(A)
}
\author{
Derek Chiu
}

