% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcNormalRange.R
\name{pcNormalRange}
\alias{pcNormalRange}
\title{Range of cluster strengths from pcNormal simulations}
\usage{
pcNormalRange(sim.dat, cl, int = 5)
}
\arguments{
\item{sim.dat}{a list object from \code{pcNormal}}

\item{cl}{cluster memberships}

\item{int}{every \code{int} data sets from median ranked \code{sim.dat}
are taken. Takes every 5th dataset by default.}
}
\value{
A list with elements
\item{ranks}{ranks of each extracted dataset}
\item{ind}{indices of range of median ranked datasets}
\item{dat}{list of simulated data cover range of median ranks}
}
\description{
Range of cluster strengths from pcNormal simulations
}
\author{
Derek Chiu
}

