% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcNormalRep.R
\name{pcNormalRep}
\alias{pcNormalRep}
\title{Representative simulation of pcNormal}
\usage{
pcNormalRep(sim.dat, cl)
}
\arguments{
\item{sim.dat}{a list object from \code{pcNormal}}

\item{cl}{cluster memberships}
}
\value{
A list with elements
\item{ind}{index of representative simulation}
\item{dat}{simulation data representation of all in pcNormal}
}
\description{
Representative simulation of pcNormal
}
\author{
Derek Chiu
}

