% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcNormalSelect.R
\name{pcNormalSelect}
\alias{pcNormalSelect}
\title{Select pcNormal simulated datasets}
\usage{
pcNormalSelect(sim.dat, cl, type = c("rep", "range"), int = 5)
}
\arguments{
\item{sim.dat}{a list object from \code{pcNormal}}

\item{cl}{cluster memberships}

\item{type}{select either the representative dataset or a range of datasets}

\item{int}{every \code{int} data sets from median ranked \code{sim.dat}
are taken. Takes every 5th dataset by default.}
}
\value{
A list with elements
\item{ranks}{When \code{type = "range"}, ranks of each extracted dataset shown}
\item{ind}{index of representative simulation}
\item{dat}{simulation data representation of all in pcNormal}
}
\description{
Select pcNormal simulated datasets
}
\author{
Derek Chiu
}

