% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclust.R
\name{print.conclust}
\alias{print.conclust}
\title{Prints terms generated by conclust}
\usage{
\method{print}{conclust}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=conclust]{conclust()}}}

\item{...}{additional arguments not used}
}
\description{
Prints terms generated by conclust
}
\keyword{conclust}
