% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silStats.R
\name{silStats}
\alias{silStats}
\title{Silhouette width statistics}
\usage{
silStats(dat, cl)
}
\arguments{
\item{dat}{data matrix}

\item{cl}{integer vector of cluster memberships}
}
\value{
A data frame with fN and aP
}
\description{
Computes the fraction of negative silhouette widths and average of positive silhouette widths.
}
\author{
Derek Chiu
}

