% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclust.R
\name{terms.conclust}
\alias{terms.conclust}
\title{Returns terms generated by conclust}
\usage{
\method{terms}{conclust}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=conclust]{conclust()}}}

\item{...}{additional arguments not used}
}
\description{
A function that returns the terms and their cosine cosimilarities
produced by \code{\link[=conclust]{conclust()}}
}
\keyword{conclust}
