% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedAlgs.R
\name{weightedAlgs}
\alias{weightedAlgs}
\title{Weighted algorithms}
\usage{
weightedAlgs(index.mat, top = 5)
}
\arguments{
\item{index.mat}{matrix of indices for each algorthim outputted
from \code{indices}.}

\item{top}{how many top algorithms to include for weighting. Defaults to 5.}
}
\value{
A matrix of weighted algorithms.
}
\description{
Obtain a subset of algorithms weighed by PAC and CHI
}
\author{
Derek Chiu
}

