% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRelabel.R
\name{classRelabel}
\alias{classRelabel}
\title{Relabel clustering categories to match to a standard}
\usage{
classRelabel(clust, cl.ref)
}
\arguments{
\item{clust}{vector of cluster assignments}

\item{cl.ref}{reference labels to match to}
}
\value{
A vector of relabeled cluster assignments
}
\description{
Relabel clustering categories to match to a standard by minimizing the minimum in Frobenius norm between the two labels.
}
\author{
Aline Talhouk
}

