% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combineAlgs.R
\name{combineAlgs}
\alias{combineAlgs}
\title{Combine the consensus results from all algorithms}
\usage{
combineAlgs(res.CC, res.CCP)
}
\arguments{
\item{res.CC}{results from using ConClust}

\item{res.CCP}{results from using ConClustPlus}
}
\value{
Either a list of all consensus matrices or a data frame
showing all the consensus classes
}
\description{
The function can output either the consensus matrices
or consensus classes for all algorithms
}

