% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conclust.R
\name{conclust}
\alias{conclust}
\title{Algorithm designed to efficiently extract keywords from a cosine similarity matrix}
\usage{
conclust(
  sim_mat,
  seed_words,
  sim_thresh = 0.25,
  max_n = 50,
  dictionary = NULL,
  exclude = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sim_mat}{A cosine similarity matrix produced by cosine.}

\item{seed_words}{A set of user-provided seed words that best represent the target concept.}

\item{sim_thresh}{Minimum cosine similarity a candidate word must have to the existing set of keywords
for it to be added.}

\item{max_n}{The maximum size of the output set of keywords.}

\item{dictionary}{An optional dictionary that maps metadata, such as definitions, to keywords.}

\item{exclude}{A vector of words that the user does not want included in the final keyword set.}

\item{verbose}{If true, conclust will produce live updates as it adds keywords.}
}
\description{
This function takes a cosine similarity matrix derived from a word embedding model,
along with a set of seed words and outputs a semantically-related set of keywords of
a length and cosimilarity determined by the user
}
\examples{
\dontrun{
# Create a set of keywords using a pre-defined set of seeds
seeds <- c("October", "November")
months <- conclust(fastText_eng_sample, seed_words = seeds, max_n = 8)
}
}
\keyword{conclust}
