% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConClust.R
\name{ConClust}
\alias{ConClust}
\title{Consensus clustering over samples and algorithms}
\usage{
ConClust(x, k, pItem = 0.8, reps = 1000, method = NULL, seed = 123456,
  seed.method = 1, dir = NULL)
}
\arguments{
\item{x}{data matrix; genes are rows and samples are columns}

\item{k}{number of clusters requested}

\item{pItem}{proportion of items to be used in subsampling within an algorithm}

\item{reps}{number of subsamples}

\item{method}{vector of clustering algorithms for performing consensus clustering. Must be
any number of the following: "nmfDiv", "nmfEucl", "hcAEucl", "hcDianaEucl", "kmEucl",
"kmSpear", "kmMI", "pamEucl", "pamSpear", "pamMI".}

\item{seed}{random seed to use for NMF-based algorithms}

\item{seed.method}{seed to use to ensure each method operates on the same set of subsamples}

\item{dir}{directory where returned object will be saved at each iteration (as an RDS object).
No output file is saved if \code{file} is \code{NULL}.}
}
\value{
An array of dimension \code{nrow(x)} by \code{reps} by \code{length(methods)}
Each slice of the array is a matrix showing consensus clustering results for
algorithms (method). The matrices have a row for each sample, and a column for each
subsample. Each entry represents a class membership.
}
\description{
Generates multiple runs for consensus clustering among replicated subsamples
of a dataset as well as across different clustering algorithms.
}
\author{
Derek Chiu, Aline Talhouk
}

