% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusClass.R
\name{consensusClass}
\alias{consensusClass}
\title{Obtain consensus classes}
\usage{
consensusClass(x, k, method = "average")
}
\arguments{
\item{x}{a consensus matrix}

\item{k}{desired number of clusters}

\item{method}{linkage type for hierarchical clustering}
}
\value{
assignments for the consensus class
}
\description{
Performs hierarchical clustering on a consensus matrix to obtain
consensus class labels.
}
\author{
Derek Chiu
}

