% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnImpute.R
\name{knnImpute}
\alias{knnImpute}
\title{Function to any missing clustering (due to resampling using K nearest neighbour)}
\usage{
knnImpute(res, tdat = tdat, seed = 123456)
}
\arguments{
\item{res}{is a vector representing a column of the output from ConClust}

\item{tdat}{is the data matrix with samples are rows and genes as columns}

\item{seed}{set random seed for reproducibility}
}
\value{
results matrix with missing values filled in except when fewer than 3 out 5 nearest neighbours are in agreement
}
\description{
Function to any missing clustering (due to resampling using K nearest neighbour)
}
\author{
Aline Talhouk
}

