% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyclust.R
\name{terms.keyclust}
\alias{terms.keyclust}
\title{Returns terms generated by keyclust}
\usage{
\method{terms}{keyclust}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=keyclust]{keyclust()}}}

\item{...}{additional arguments not used}
}
\description{
A function that returns the terms and their cosine cosimilarities
produced by \code{\link[=keyclust]{keyclust()}}
}
\keyword{keyclust}
